+++ src/vdp_builtins.cpp
@@ -4,7 +4,7 @@ namespace VDP {
 namespace Schema {
 Motor::Motor(std::string name, vex::motor &mot)
     : Record(name), mot(mot), pos(new Double("Position(deg)")),
-      vel(new Double("velocity(dps)")), temp(new Double("Temperature(C)")),
+      vel(new Double("velocity(dps)")), temp(new Uint8("Temperature(C)")),
       voltage(new Double("Voltage(V)")), current(new Double("Current(%)")) {
   Record::setFields({pos, vel, temp, voltage, current});
 }
@@ -12,7 +12,7 @@ Motor::Motor(std::string name, vex::motor &mot)
 void Motor::fetch() {
   pos->setValue(mot.position(vex::rotationUnits::deg));
   vel->setValue(mot.velocity(vex::velocityUnits::dps));
-  temp->setValue(mot.temperature(vex::temperatureUnits::celsius));
+  temp->setValue((uint8_t)mot.temperature(vex::temperatureUnits::celsius));
   voltage->setValue(mot.voltage(vex::voltageUnits::volt));
   current->setValue(mot.current(vex::percentUnits::pct));
 }
